% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{accessor}
\alias{accessor}
\alias{[,Domain-method}
\alias{$,Domain-method}
\alias{$<-,Domain-method}
\title{Accessor}
\usage{
\S4method{[}{Domain}(x, i, j)

\S4method{$}{Domain}(x, name)

\S4method{$}{Domain}(x, name) <- value
}
\arguments{
\item{x}{(\code{Domain}).}

\item{i, j}{(\code{numeric} or \code{character})\cr indices of the columns/rows to
select.}

\item{name}{(\code{character})\cr name of the object in \code{x} to access.}

\item{value}{(\code{atomic} or \code{domain}).}
}
\description{
Extract data from a \code{Domain} object.
}
\section{Functions}{
\itemize{
\item \code{[,Domain-method}: select with brackets and indices.

\item \code{$,Domain-method}: select a variable with the dollar symbol \code{$}.

\item \code{$<-,Domain-method}: assign a variable with the dollar symbol \code{$}.
}}

\examples{
dta <- DM(
  dm[1:3, c("STUDYID", "SUBJID", "RACE")],
  data.frame(row.names = "STUDYID", label = "Study Identifier")
)
dta
dta[1, ]
dta[, 1]
dta[1, 1]
dta[1:2, ]
dta[, 1:2]
dta[1:2, 1:2]

dta <- DM(
  dm[1:2, c("STUDYID", "SUBJID", "RACE", "AGE")],
  data.frame(row.names = "STUDYID", label = "Study Identifier")
)
dta
dta$RACE
dta$STUDYID


dta_1 <- DM(
  dm[1:3, c("STUDYID", "SUBJID", "ARMCD")],
  data.frame(
    row.names = "STUDYID",
    label = "Study Identifier",
    comment = "std"
  )
)

dta_1$new_var <- 1:3
dta_1

dta_2 <- DM(
  dm[1:3, c("STUDYID", "SUBJID", "AGE")],
  data.frame(
    row.names = c("STUDYID", "AGE"),
    names = c("Study Identifier", "Age"),
    label = c(NA, "Age")
  )
)
dta_2$AGE
dta_1$age <- dta_2$AGE
dta_1

}
