% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sdtm.R
\docType{data}
\name{data-sdtm}
\alias{data-sdtm}
\alias{ae}
\alias{dm}
\alias{eg}
\alias{lb}
\alias{mh}
\alias{pe}
\alias{vs}
\title{SDTM - Sample Study Dataset}
\format{
SDTM

\itemize{
\item \code{ae}: 18 Variables, 16 Observations.
}

\itemize{
\item \code{dm}: 16 Variables, 5 Observations.
}

\itemize{
\item \code{eg}: 21 Variables, 56 Observations.
}

\itemize{
\item \code{lb}: 29 Variables, 83 Observations.
}

\itemize{
\item \code{mh}: 12 Variables, 18 Observations.
}

\itemize{
\item \code{pe}: 13 Variables, 65 Observations.
}

\itemize{
\item \code{vs}: 18 Variables, 72 Observations.
}
}
\source{
\insertCite{Emam2019;textual}{ClinAD}.
}
\usage{
ae

dm

eg

lb

mh

pe

vs
}
\description{
A sample study dataset following SDTM guidelines.
This generally follows a controlled terminolgy \insertCite{NCI2020}{ClinAD}.
}
\section{Functions}{
\itemize{
\item \code{ae}: Adverse Events data.

\item \code{dm}: Subject demographics data.

\item \code{eg}: Electrogardiogram data.

\item \code{lb}: Laboratory findings data.

\item \code{mh}: Medical History data.

\item \code{pe}: Physical examination data.

\item \code{vs}: Vital Signs data..
}}

\examples{
head(ae)

dm

head(eg)

head(lb)

head(mh)

head(pe)

head(vs)

}
\references{
\insertAllCited{}
}
\keyword{datasets}
