% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R, R/class_dm.R
\docType{class}
\name{domain}
\alias{domain}
\alias{Domain}
\alias{Domain,data.frame,missing-method}
\alias{Domain,data.frame,data.frame-method}
\alias{DM-class}
\alias{DM}
\alias{DM,data.frame,missing-method}
\alias{DM,data.frame,data.frame-method}
\title{S4 Class: domain}
\usage{
Domain(df, columns)

\S4method{Domain}{data.frame,missing}(df, columns = NULL)

\S4method{Domain}{data.frame,data.frame}(df, columns)

DM(df, columns)

\S4method{DM}{data.frame,missing}(df, columns = NULL)

\S4method{DM}{data.frame,data.frame}(df, columns)
}
\arguments{
\item{df}{(\code{data.frame}).}

\item{columns}{(\code{data.frame}).}
}
\description{
S4 Class: domain
}
\section{Functions}{
\itemize{
\item \code{Domain}: generic Domain

\item \code{Domain,data.frame,missing-method}: create Domain with missing column description.

\item \code{Domain,data.frame,data.frame-method}: create Domain with column description.

\item \code{DM-class}: class DM

\item \code{DM}: generic DM

\item \code{DM,data.frame,missing-method}: create DM with missing column description.

\item \code{DM,data.frame,data.frame-method}: create DM with column description.
}}

\section{Slots}{

\describe{
\item{\code{df}}{(\code{data.frame})\cr the data.}

\item{\code{columns}}{(\code{data.frame})\cr metadata, information about columns.}
}}

\examples{
Domain(df = dm)


columns <- data.frame(
  row.names = c("STUDYID", "AGE"),
  label = c("Study Identifier", "Age")
)
Domain(df = dm, columns = columns)

DM(df = dm)


columns <- data.frame(
  row.names = c("STUDYID", "AGE"),
  label = c("Study Identifier", "Age")
)
DM(df = dm, columns = columns)

}
