% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_columns.R
\name{expand_columns}
\alias{expand_columns}
\title{Expand Columns}
\usage{
expand_columns(x, y)
}
\arguments{
\item{x, y}{(\code{data.frame})}
}
\value{
A \code{data.frame} with the columns from the two datasets.
}
\description{
Row-binds to dataset while insuring column match.
}
\details{
The columns metadata must follow the column, even when transfered
from a dataset to another one. However, while the \code{data.frame} format with
row number equal to data column number, it is not garranty the two data
sets have the same metadata composition. This function, makes sure of the
column match between two datasets before row binding.
}
\examples{

t1 <- data.frame(
  row.names = "STUDYID", label = "Study Identifier", names = "name"
)
# |        |label            |names |
# |:-------|:----------------|:-----|
# |STUDYID |Study Identifier |name  |
t2 <- data.frame(
  row.names = c("SUBJID", "AGE"),
  names = c("Subject Identifier", "Age")
)

# |       |names              |
# |:------|:------------------|
# |SUBJID |Subject Identifier |
# |AGE    |Age                |

expand_columns(t1, t2)
# |        |label            |names              |
# |:-------|:----------------|:------------------|
# |STUDYID |Study Identifier |name               |
# |SUBJID  |NA               |Subject Identifier |
# |AGE     |NA               |Age                |

}
